feat(backend): switched calls to pokeapi to asyncronous
split logic.py into more managable individual files
