+++ backend/app/logic/get_types.py
@@ -1,25 +1,22 @@
-import httpx
-from typing_extensions import TypedDict
-from glom import Coalesce, glom
-from app.pokeapi_requests import PokemonList, RawType, get_all_types, get_type
 import asyncio
 
+import httpx
+from glom import Coalesce, glom
 
-class TypesType(TypedDict):
-    name: str
-    img: str
+from app.pokeapi_requests import PokemonList, RawType, get_all_types, get_type
+from app.types import TypesType
 
 
 async def get_types(client: httpx.AsyncClient) -> list[TypesType]:
     pokemon_list: PokemonList = await get_all_types(client)
-    type_list: list[str] = [species['name'] for species in pokemon_list.get('results', {})]
-    raw_type_task_list: list[RawType] = []
+    type_list: list[str] = [species.get('name') for species in pokemon_list.get('results', {})]
+    raw_type_task_list: list[asyncio.Task[RawType]] = []
     result: list[TypesType] = []
 
     async with asyncio.TaskGroup() as tg:
         for type in type_list:
-            raw_type: RawType = tg.create_task(get_type(type, client))
-            raw_type_task_list.append(raw_type)
+            task: asyncio.Task[RawType] = tg.create_task(get_type(type, client))
+            raw_type_task_list.append(task)
 
     for raw_type in raw_type_task_list:
         img: str | None = glom(
@@ -35,5 +32,5 @@ async def get_types(client: httpx.AsyncClient) -> list[TypesType]:
             ),
         )
         if img:
-            result.append({'name': raw_type.result()['name'], 'img': img})
+            result.append({'name': raw_type.result().get('name', ''), 'img': img})
     return result
